// src/app/page/master/map-jenispegawai-departemen/map-jenispegawai-departemen.interface.ts
import { EventEmitter } from '@angular/core';

export interface MapJenispegawaiDepartemen {
	kdProduk?: string;
	kdEvent?: number;
	label?: any;
	value?: any;
	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
}